Changed pinstall adding a condition if the icewm template exists - see previous icewm removal commits

+++ woof-code/rootfs-skeleton/pinstall.sh
@@ -9,6 +9,7 @@
 #120225 create symlink release-notes.htm to actual release-notes file. NO.
 #120225 backup doc files, refer /usr/sbin/indexgen.sh.
 #120818 now have /etc/xdg in Woof, taken out of xdg_puppy PET, relocated pinstall.sh to here.
+#132211 removed icewm template from default, who uses that anyway?
 
 WKGDIR="`pwd`"
 
@@ -60,30 +61,31 @@ echo -n "vmlinuz" > /tmp/vmlinuzforpup.txt
 
 #120818 now have /etc/xdg in Woof, taken out of xdg_puppy PET, relocated pinstall.sh to here...
 #this code is to fix the icewm menu...
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name evilwm`" = "" ];then
- grep -v ' evilwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+if [ -f ./etc/xdg/templates/_root_.icewm_menu ];then
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name evilwm`" = "" ];then
+  grep -v ' evilwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name fluxbox`" = "" ];then
+  grep -v ' fluxbox$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name fvwm95`" = "" ];then
+  grep -v ' fvwm95$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name jwm`" = "" ];then
+  grep -v ' jwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name pwm`" = "" ];then
+  grep -v ' pwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
+ if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name xfce4-session`" = "" ];then
+  grep -v ' xfce4-session$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
+  mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
+ fi
 fi
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name fluxbox`" = "" ];then
- grep -v ' fluxbox$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
-fi
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name fvwm95`" = "" ];then
- grep -v ' fvwm95$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
-fi
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name jwm`" = "" ];then
- grep -v ' jwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
-fi
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name pwm`" = "" ];then
- grep -v ' pwm$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
-fi
-if [ "`find ./usr/local/bin ./usr/bin ./usr/sbin ./usr/X11R7/bin -maxdepth 1 -type f -name xfce4-session`" = "" ];then
- grep -v ' xfce4-session$' ./etc/xdg/templates/_root_.icewm_menu > /tmp/_root_.icewm_menu
- mv -f /tmp/_root_.icewm_menu ./etc/xdg/templates/
-fi
-
 
 #end#
